Move classifcation and regression into seperate files
