+++ linfa-svm/src/solver_smo.rs
@@ -1,6 +1,5 @@
 use super::{ExitReason, Float, SolverParams, SvmResult};
-use super::permutable_kernel::{Permutable, PermutableKernel, PermutableKernelOneClass, PermutableKernelRegression};
-use linfa_kernel::Kernel;
+use super::permutable_kernel::Permutable;
 
 /// Status of alpha variables of the solver
 #[derive(Debug)]
@@ -839,372 +838,5 @@ impl<'a, A: Float, K: Permutable<'a, A>> SolverState<'a, A, K> {
             iterations: iter,
             kernel: self.kernel.inner(),
         }
-    }
-}
-
-pub struct Classification;
-
-impl Classification {
-    pub fn fit_c<'a, A: Float>(
-        params: &'a SolverParams<A>,
-        kernel: &'a Kernel<A>,
-        targets: &'a [bool],
-        cpos: A,
-        cneg: A,
-    ) -> SvmResult<'a, A> {
-        let bounds = targets
-            .iter()
-            .map(|x| if *x { cpos } else { cneg })
-            .collect::<Vec<_>>();
-
-        let kernel = PermutableKernel::new(kernel, targets.to_vec());
-
-        let solver = SolverState::new(
-            vec![A::zero(); targets.len()],
-            vec![-A::one(); targets.len()],
-            targets.to_vec(),
-            kernel,
-            bounds,
-            params,
-            false,
-        );
-
-        let mut res = solver.solve();
-
-        res.alpha = res
-            .alpha
-            .into_iter()
-            .zip(targets.iter())
-            .map(|(a, b)| if *b { a } else { -a })
-            .collect();
-
-        res
-    }
-
-    pub fn fit_nu<'a, A: Float>(
-        params: &'a SolverParams<A>,
-        kernel: &'a Kernel<A>,
-        targets: &'a [bool],
-        nu: A,
-    ) -> SvmResult<'a, A> {
-        let mut sum_pos = nu * A::from(targets.len()).unwrap() / A::from(2.0).unwrap();
-        let mut sum_neg = nu * A::from(targets.len()).unwrap() / A::from(2.0).unwrap();
-        let init_alpha = targets
-            .iter()
-            .map(|x| {
-                if *x {
-                    let val = A::min(A::one(), sum_pos);
-                    sum_pos -= val;
-                    val
-                } else {
-                    let val = A::min(A::one(), sum_neg);
-                    sum_neg -= val;
-                    val
-                }
-            })
-            .collect::<Vec<_>>();
-
-        let kernel = PermutableKernel::new(kernel, targets.to_vec());
-
-        let solver = SolverState::new(
-            init_alpha,
-            vec![A::zero(); targets.len()],
-            targets.to_vec(),
-            kernel,
-            vec![A::one(); targets.len()],
-            params,
-            true,
-        );
-
-        let mut res = solver.solve();
-
-        let r = res.r.unwrap();
-
-        res.alpha = res
-            .alpha
-            .into_iter()
-            .zip(targets.iter())
-            .map(|(a, b)| if *b { a } else { -a })
-            .map(|x| x / r)
-            .collect();
-        res.rho /= r;
-        res.obj /= r * r;
-
-        res
-    }
-
-    pub fn fit_one_class<'a, A: Float + num_traits::ToPrimitive>(
-        params: &'a SolverParams<A>,
-        kernel: &'a Kernel<A>,
-        nu: A,
-    ) -> SvmResult<'a, A> {
-        let size = kernel.size();
-        let n = (nu * A::from(size).unwrap()).to_usize().unwrap();
-
-        let init_alpha = (0..size)
-            .map(|x| {
-                if x < n {
-                    A::one()
-                } else if x == n {
-                    nu * A::from(size).unwrap() - A::from(x).unwrap()
-                } else {
-                    A::zero()
-                }
-            })
-            .collect::<Vec<_>>();
-
-        let kernel = PermutableKernelOneClass::new(kernel);
-
-        let solver = SolverState::new(
-            init_alpha,
-            vec![A::zero(); size],
-            vec![true; size],
-            kernel,
-            vec![A::one(); size],
-            params,
-            false,
-        );
-
-        solver.solve()
-    }
-}
-
-pub struct Regression;
-
-impl Regression {
-    pub fn fit_epsilon<'a, A: Float>(
-        params: &'a SolverParams<A>,
-        kernel: &'a Kernel<A>,
-        target: &'a [A],
-        c: A,
-        p: A,
-    ) -> SvmResult<'a, A> {
-        let mut linear_term = vec![A::zero(); 2 * target.len()];
-        let mut targets = vec![true; 2 * target.len()];
-
-        for i in 0..target.len() {
-            linear_term[i] = p - target[i];
-            targets[i] = true;
-
-            linear_term[i + target.len()] = p + target[i];
-            targets[i] = false;
-        }
-
-        let kernel = PermutableKernelRegression::new(kernel);
-        let solver = SolverState::new(
-            vec![A::zero(); 2 * target.len()],
-            linear_term,
-            targets.to_vec(),
-            kernel,
-            vec![c; target.len()],
-            params,
-            false,
-        );
-
-        let mut res = solver.solve();
-
-        for i in 0..target.len() {
-            let tmp = res.alpha[i+target.len()];
-            res.alpha[i] -= tmp;
-        }
-        res.alpha.truncate(target.len());
-
-        res
-
-
-    }
-
-    pub fn fit_nu<'a, A: Float>(
-        params: &'a SolverParams<A>,
-        kernel: &'a Kernel<A>,
-        target: &'a [A],
-        c: A,
-        nu: A,
-    ) -> SvmResult<'a, A> {
-        let mut alpha = vec![A::zero(); 2 * target.len()];
-        let mut linear_term = vec![A::zero(); 2 * target.len()];
-        let mut targets = vec![true; 2 * target.len()];
-
-        let mut sum = c * nu * A::from(target.len()).unwrap() / A::from(2.0).unwrap();
-        for i in 0..target.len() {
-            alpha[i] = A::min(sum, c);
-            alpha[i + target.len()] = A::min(sum, c);
-            sum -= alpha[i];
-
-            linear_term[i] = -target[i];
-            targets[i] = true;
-
-            linear_term[i + target.len()] = target[i];
-            targets[i] = false;
-        }
-
-        let kernel = PermutableKernelRegression::new(kernel);
-        let solver = SolverState::new(
-            alpha,
-            linear_term,
-            targets.to_vec(),
-            kernel,
-            vec![c; target.len()],
-            params,
-            false,
-        );
-
-        solver.solve()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{Classification, Regression, Permutable, PermutableKernel, SolverParams};
-    use linfa::metrics::IntoConfusionMatrix;
-    use linfa_kernel::Kernel;
-    use ndarray::{array, Array, Axis};
-    use ndarray_rand::rand_distr::Uniform;
-    use ndarray_rand::RandomExt;
-
-    #[test]
-    fn test_swappable_kernel() {
-        let dist = array![[1.0, 0.3, 0.1], [0.3, 1.0, 0.5], [0.1, 0.5, 1.0]];
-        let targets = vec![true, true, true];
-        let dist = Kernel::from_dense(dist);
-        let mut kernel = PermutableKernel::new(&dist, targets);
-
-        assert_eq!(kernel.distances(0, 3), &[1.0, 0.3, 0.1]);
-        assert_eq!(kernel.distances(1, 3), &[0.3, 1.0, 0.5]);
-        assert_eq!(kernel.distances(2, 3), &[0.1, 0.5, 1.0]);
-
-        // swap first two nodes
-        kernel.swap_indices(0, 1);
-
-        assert_eq!(kernel.distances(0, 3), &[1.0, 0.3, 0.5]);
-        assert_eq!(kernel.distances(1, 3), &[0.3, 1.0, 0.1]);
-        assert_eq!(kernel.distances(2, 3), &[0.5, 0.1, 1.0]);
-
-        // swap second and third node
-        kernel.swap_indices(1, 2);
-
-        assert_eq!(kernel.distances(0, 3), &[1.0, 0.5, 0.3]);
-        assert_eq!(kernel.distances(1, 3), &[0.5, 1.0, 0.1]);
-        assert_eq!(kernel.distances(2, 3), &[0.3, 0.1, 1.0]);
-    }
-
-    #[test]
-    fn test_c_classification() {
-        // generate two clusters with 100 samples each
-        let entries = ndarray::stack(
-            Axis(0),
-            &[
-                Array::random((10, 2), Uniform::new(-10., -5.)).view(),
-                Array::random((10, 2), Uniform::new(5., 10.)).view(),
-            ],
-        )
-        .unwrap();
-        let targets = (0..20).map(|x| x < 10).collect::<Vec<_>>();
-
-        let kernel = Kernel::gaussian(entries, 100.);
-        let params = SolverParams {
-            eps: 1e-3,
-            shrinking: false,
-        };
-
-        let svc = Classification::fit_c(&params, &kernel, &targets, 1.0, 1.0);
-        println!("{}", svc);
-
-        let pred = kernel
-            .dataset
-            .outer_iter()
-            .map(|x| svc.predict(x))
-            .map(|x| x > 0.0)
-            .collect::<Vec<_>>();
-
-        let cm = pred.into_confusion_matrix(&targets);
-        assert_eq!(cm.accuracy(), 1.0);
-    }
-
-    #[test]
-    fn test_nu_classification() {
-        // generate two clusters with 100 samples each
-        let entries = ndarray::stack(
-            Axis(0),
-            &[
-                Array::random((10, 2), Uniform::new(-10., -5.)).view(),
-                Array::random((10, 2), Uniform::new(5., 10.)).view(),
-            ],
-        )
-        .unwrap();
-        let targets = (0..20).map(|x| x < 10).collect::<Vec<_>>();
-
-        let kernel = Kernel::gaussian(entries, 100.);
-        let params = SolverParams {
-            eps: 1e-1,
-            shrinking: false,
-        };
-
-        let svc = Classification::fit_nu(&params, &kernel, &targets, 0.1);
-        println!("{}", svc);
-
-        let pred = kernel
-            .dataset
-            .outer_iter()
-            .map(|x| svc.predict(x))
-            .map(|x| x > 0.0)
-            .collect::<Vec<_>>();
-
-        dbg!(&svc.alpha);
-        let cm = pred.into_confusion_matrix(&targets);
-        assert_eq!(cm.accuracy(), 1.0);
-    }
-
-    #[test]
-    fn test_reject_classification() {
-        // generate two clusters with 100 samples each
-        let entries = Array::random((100, 2), Uniform::new(-4., 4.));
-        let kernel = Kernel::gaussian(entries, 100.);
-
-        let params = SolverParams {
-            eps: 1e-3,
-            shrinking: false,
-        };
-
-        let svc = Classification::fit_one_class(&params, &kernel, 0.1);
-        println!("{}", svc);
-
-        // now test that points outside the circle are rejected
-        let validation = Array::random((100, 2), Uniform::new(-10., 10f32));
-        let pred = validation
-            .outer_iter()
-            .map(|x| svc.predict(x) > 0.0)
-            .collect::<Vec<_>>();
-
-        // count the number of correctly rejected samples
-        let mut rejected = 0;
-        let mut total = 0;
-        for (pred, pos) in pred.iter().zip(validation.outer_iter()) {
-            let distance = (pos[0]*pos[0] + pos[1]*pos[1]).sqrt();
-            if distance >= 5.0 {
-                if !pred {
-                    rejected += 1;
-                }
-                total += 1;
-            }
-        }
-
-        // at least 95% should be correctly rejected
-        assert!((rejected as f32) / (total as f32) > 0.95);
-    }
-
-    #[test]
-    fn test_linear_epsilon_regression() {
-        let target = Array::linspace(0., 10., 100);
-        let entries = Array::ones((100, 2));
-        let kernel = Kernel::gaussian(entries, 100.);
-
-        let params = SolverParams {
-            eps: 1e-3,
-            shrinking: false
-        };
-
-        let svr = Regression::fit_epsilon(&params, &kernel, &target.as_slice().unwrap(), 1.0, 0.1);
-        println!("{}", svr);
     }
 }
